Add input test after install driver via installer
