+++ qemu/tests/cfg/win_virtio_driver_install_by_installer.cfg
@@ -210,6 +210,15 @@
                     balloon = balloon0
                     balloon_dev_devid = balloon0
                     balloon_dev_add_bus = yes
+                - with_vioinput:
+                    driver_name = vioinput
+                    device_name = VirtIO Input Driver
+                    device_hwid = '"PCI\\VEN_1AF4&DEV_1052"'
+                    driver_test_names = 'vioinput'
+                    inputs = input1
+                    input_dev_bus_type_input1 = virtio
+                    input_dev_type_input1 = keyboard
+                    key_table_file = key_to_keycode_win.json
     variants:
         - installer_version_check:
             only all_drivers
